- Merged 2 calculate age methods to one - In /botinfo, replaced the name field with field with the mention of the bot - Created StringUtils.booleanToEmoji - Changed the type of Property.estateTax to BigInteger - Currency amounts in /property list are now formatted with StringUtils.numberFormat - Made Rent use BigInteger - Added :x:'s to the fail messages of /reports - In /shop view, the price is now formatted with StringUtils.numberFormat - Made ShopItem use BigInteger

+++ src/main/java/dev/darealturtywurty/superturtybot/modules/economy/Property.java
@@ -23,7 +23,7 @@ public class Property {
     private long buyDate;
 
     private List<Long> previousOwners;
-    private int estateTax;
+    private BigInteger estateTax;
     private int upgradeLevel;
     
     private Property(Builder builder) {
@@ -92,14 +92,14 @@ public class Property {
         private final BigInteger price;
         private final List<BigInteger> upgradePrices = new ArrayList<>();
         private final List<Long> previousOwners = new ArrayList<>();
-        private final int estateTax;
+        private final BigInteger estateTax;
 
         private long owner = -1;
         private Rent rent;
         private Loan mortgage;
         private int upgradeLevel = 0;
 
-        public Builder(String name, String description, BigInteger price, int estateTax) {
+        public Builder(String name, String description, BigInteger price, BigInteger estateTax) {
             this.name = name;
             this.description = description;
             this.price = price;
